CLeanup of db save messages and finalised project list view

+++ app/Repositories/UserRepository.php
@@ -14,48 +14,53 @@ class UserRepository
   protected $user;
 
   public function __construct(User $user)
-	{
-		$this->user = $user;
-	}
+  {
+    $this->user = $user;
+  }
 
   public function getById($id)
-	{
-		return $this->user->findOrFail($id);
-	}
-
-  public function createIfNotFound(Array $inputs)
   {
-    $user = $this->user->where('name', $inputs['name'])->first();
+    return $this->user->findOrFail($id);
+  }
 
-    if (!isset($user)){
-        $user = new $this->user;
-        return $this->save($user, $inputs);
-    }
-    else {
-      return $user;
-    }
+  public function create(Array $inputs)
+  {
+    $user = new $this->user;
+    return $this->save($user, $inputs);
   }
 
   public function update($id, Array $inputs)
-	{
-		$this->save($this->getById($id), $inputs);
-	}
+  {
+    return $this->save($this->getById($id), $inputs);
+  }
 
   public function update_password($id, Array $inputs)
-	{
+  {
+    $result = new \stdClass();
+    $result->result = 'success';
+    $result->msg = '';
+
     $user = $this->getById($id);
     if (isset($inputs['password']) && trim($inputs['password']) != ''){
       $user->password = bcrypt($inputs['password']);
+      $user->save();
+      $result->msg = 'Password saved';
+    } else {
+      $result->result = 'error';
+      $result->msg = 'Password not saved';
     }
-		$user->save();
-	}
+    return $result;
+  }
 
-	private function save(User $user, Array $inputs)
-	{
+  private function save(User $user, Array $inputs)
+  {
+    $result = new \stdClass();
+    $result->result = 'success';
+    $result->msg = '';
 
     // Required fields
-		$user->name = $inputs['name'];
-    $user->email = $inputs['email'];
+    if (isset($inputs['name'])) {$user->name = $inputs['name'];}
+    if (isset($inputs['email'])) {$user->email = $inputs['email'];}
     // Password special case
     if (isset($inputs['password']) && trim($inputs['password']) != ''){
       $user->password = bcrypt($inputs['password']);
@@ -69,10 +74,17 @@ class UserRepository
     if (isset($inputs['job_role'])) {$user->job_role = $inputs['job_role'];}
     if (isset($inputs['employee_type'])) {$user->employee_type = $inputs['employee_type'];}
     // Boolean
-    if (isset($inputs['is_manager'])) {$user->is_manager = $inputs['is_manager'];}
-    if (isset($inputs['from_otl'])) {$user->from_otl = $inputs['from_otl'];}
+    $user->is_manager = isset($inputs['is_manager']);
+    $user->from_otl = isset($inputs['from_otl']);
 
-    $user->save();
+    try {
+      $user->save();
+    }
+    catch (\Illuminate\Database\QueryException $ex){
+        $result->result = 'error';
+        $result->msg = 'Message:</BR>'.$ex->getMessage();
+        return $result;
+    }
 
     // Now we have to treat the manager
     if ($inputs['manager_id'] != -1){
@@ -94,21 +106,42 @@ class UserRepository
 
     // Now we need to save the roles
     if (Entrust::can('role-assign')){
-      DB::table('role_user')->where('user_id',$user->id)->delete();
-      foreach ($inputs['roles'] as $key => $value) {
-              $user->attachRole($value);
-          }
+      if (isset($inputs['roles'])) {
+        DB::table('role_user')->where('user_id',$user->id)->delete();
+        foreach ($inputs['roles'] as $key => $value) {
+          $user->attachRole($value);
+        }
+      }
     }
-    return $user;
-	}
 
-	public function destroy($id)
-	{
+    $result->msg = 'User '.$user->name.' saved successfully.';
+
+    return $result;
+  }
+
+  public function destroy($id)
+  {
+    $result = new \stdClass();
+    $result->result = 'success';
+    $result->msg = '';
+
     $user = $this->getById($id);
-    $user->managers()->detach();
-		$this->getById($id)->delete();
-    return 'success';
-	}
+    $name = $user->name;
+
+    try {
+      $user->managers()->detach();
+      $user->delete();
+
+    } catch (\Illuminate\Database\QueryException $ex){
+        $result->result = 'error';
+        $result->msg = '</BR>Message:</BR>'.$ex->getMessage();
+        return $result;
+    }
+
+    $result->msg = 'User '.$name.' deleted successfully.';
+
+    return $result;
+  }
 
   public function getListOfUsers()
   {
@@ -118,23 +151,23 @@ class UserRepository
     *   Then we will need to use in the view page the name of the table.column. This is so that it knows how to do proper sorting or search.
     **/
     $userList = DB::table('users')
-      ->select( 'users.id', 'users.name','users.email','users.is_manager', 'users.region',
-                'users.country', 'users.domain', 'users.management_code', 'users.job_role',
-                'users.employee_type','users_users.manager_id','u2.name AS manager_name')
-      ->leftjoin('users_users', 'users.id', '=', 'users_users.user_id')
-      ->leftjoin('users AS u2', 'u2.id', '=', 'users_users.manager_id');
+    ->select( 'users.id', 'users.name','users.email','users.is_manager', 'users.region',
+    'users.country', 'users.domain', 'users.management_code', 'users.job_role',
+    'users.employee_type','users_users.manager_id','u2.name AS manager_name')
+    ->leftjoin('users_users', 'users.id', '=', 'users_users.user_id')
+    ->leftjoin('users AS u2', 'u2.id', '=', 'users_users.manager_id');
     $data = Datatables::of($userList)->make(true);
     return $data;
   }
 
   public function getMyManagersList($id)
-	{
+  {
     $data = $this->user->findOrFail($id)->managers()->select('manager_id','name')->get();
     return $data;
-	}
+  }
 
   public function getManagersList()
-	{
+  {
     return $this->user->where('is_manager', '=','1')->lists('name','id');
-	}
+  }
 }
